831: Extracting ApiClientJwkSetService which implements the logic to get the JWKS for an ApiClient
Updating FetchApiClientJwksFilter to use ApiClientJwkSetService

https://github.com/SecureApiGateway/SecureApiGateway/issues/842
