+++ secure-api-gateway-ig-extensions/src/test/java/com/forgerock/sapi/gateway/jwks/FetchApiClientJwksFilterTest.java
@@ -15,30 +15,28 @@
  */
 package com.forgerock.sapi.gateway.jwks;
 
+import static com.forgerock.sapi.gateway.jwks.DefaultApiClientJwkSetServiceTest.createApiClientWithJwksUri;
+import static com.forgerock.sapi.gateway.jwks.DefaultApiClientJwkSetServiceTest.createJwkSet;
 import static org.forgerock.json.JsonValue.field;
 import static org.forgerock.json.JsonValue.json;
 import static org.forgerock.json.JsonValue.object;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-import java.net.URI;
 import java.net.URL;
-import java.util.List;
 import java.util.Map;
-import java.util.UUID;
+import java.util.concurrent.ExecutionException;
 import java.util.concurrent.TimeUnit;
+import java.util.concurrent.TimeoutException;
 
 import org.forgerock.http.protocol.Request;
 import org.forgerock.http.protocol.Response;
 import org.forgerock.http.protocol.Status;
 import org.forgerock.json.JsonValue;
+import org.forgerock.json.jose.exceptions.FailedToLoadJWKException;
 import org.forgerock.json.jose.jwk.JWKSet;
-import org.forgerock.json.jose.jws.JwsHeader;
-import org.forgerock.json.jose.jws.SignedJwt;
-import org.forgerock.json.jose.jwt.JwtClaimsSet;
 import org.forgerock.openig.heap.HeapException;
 import org.forgerock.openig.heap.HeapImpl;
 import org.forgerock.openig.heap.Name;
@@ -47,55 +45,38 @@ import org.forgerock.services.context.Context;
 import org.forgerock.services.context.RootContext;
 import org.forgerock.util.promise.NeverThrowsException;
 import org.forgerock.util.promise.Promise;
+import org.forgerock.util.promise.Promises;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 
-import com.forgerock.sapi.gateway.dcr.models.ApiClient;
 import com.forgerock.sapi.gateway.dcr.idm.FetchApiClientFilter;
+import com.forgerock.sapi.gateway.dcr.models.ApiClient;
 import com.forgerock.sapi.gateway.jwks.FetchApiClientJwksFilter.Heaplet;
-import com.forgerock.sapi.gateway.jwks.cache.BaseCachingJwkSetServiceTest.ReturnsErrorsJwkStore;
 import com.forgerock.sapi.gateway.jwks.mocks.MockJwkSetService;
 import com.forgerock.sapi.gateway.trusteddirectories.FetchTrustedDirectoryFilter;
 import com.forgerock.sapi.gateway.trusteddirectories.TrustedDirectory;
 import com.forgerock.sapi.gateway.trusteddirectories.TrustedDirectoryOpenBankingTest;
-import com.forgerock.sapi.gateway.trusteddirectories.TrustedDirectorySecureApiGateway;
 import com.forgerock.sapi.gateway.util.TestHandlers.TestSuccessResponseHandler;
 
 class FetchApiClientJwksFilterTest {
 
     @Test
-    void fetchJwkSetFromJwksUri() throws Exception {
+    void testFetchApiClientJwks() throws Exception {
         final JWKSet jwkSet = createJwkSet();
-        final URL jwksUri = new URL("https://directory.com/jwks/12345");
-        final MockJwkSetService jwkSetService = new MockJwkSetService(Map.of(jwksUri, jwkSet));
-        final FetchApiClientJwksFilter filter = new FetchApiClientJwksFilter(jwkSetService);
+        final MockApiClientJwkSetService apiClientJwkSetService = new MockApiClientJwkSetService(jwkSet);
+        final FetchApiClientJwksFilter filter = new FetchApiClientJwksFilter(apiClientJwkSetService);
 
-        fetchJwkSetFromJwksUri(jwkSet, jwksUri, filter);
-    }
+        final Context context = new AttributesContext(new RootContext());
+        addApiClientToContext(context, new ApiClient());
+        addTrustedDirectoryToContext(context, new TrustedDirectoryOpenBankingTest());
+        final TestSuccessResponseHandler responseHandler = new TestSuccessResponseHandler();
+        final Promise<Response, NeverThrowsException> responsePromise = filter.filter(context, new Request(), responseHandler);
+        final Response response = responsePromise.get(1, TimeUnit.MILLISECONDS);
+        assertEquals(Status.OK, response.getStatus());
+        assertTrue(responseHandler.hasBeenInteractedWith());
 
-    private void fetchJwkSetFromJwksUri(JWKSet expectedJwkSet, URL jwksUri, FetchApiClientJwksFilter filter) throws Exception {
-        final ApiClient apiClient = createApiClientWithJwksUri(jwksUri.toURI());
-        // OB Trusted Dir uses the jwksUri
-        final TrustedDirectory trustedDirectory = new TrustedDirectoryOpenBankingTest();
-        invokeFilterAndValidateSuccessResponse(expectedJwkSet, apiClient, trustedDirectory, filter);
-    }
-
-    @Test
-    void fetchJwkSetFromSoftwareStatement() throws Exception {
-        // Never expect the JwkSetService to get called in this case
-        final ReturnsErrorsJwkStore errorsJwkStore = new ReturnsErrorsJwkStore();
-        final FetchApiClientJwksFilter filter = new FetchApiClientJwksFilter(errorsJwkStore);
-        fetchJwkSetFromSoftwareStatement(filter);
-    }
-
-    private void fetchJwkSetFromSoftwareStatement(FetchApiClientJwksFilter filter) throws Exception {
-        final JWKSet jwkSet = createJwkSet();
-        // SAPI-G directory uses the software statement jwks
-        final URL secureApiGatewayJwksURI = new URL("https://blah.com");
-        final TrustedDirectory trustedDirectory = new TrustedDirectorySecureApiGateway(secureApiGatewayJwksURI);
-        final ApiClient apiClient = createApiClientWithSoftwareStatementJwks(jwkSet, trustedDirectory.getSoftwareStatementJwksClaimName());
-
-        invokeFilterAndValidateSuccessResponse(jwkSet, apiClient, trustedDirectory, filter);
+        // Verify JWKS is available in the context
+        assertEquals(jwkSet, FetchApiClientJwksFilter.getApiClientJwkSetFromContext(context));
     }
 
     @Test
@@ -103,7 +84,7 @@ class FetchApiClientJwksFilterTest {
         final TestSuccessResponseHandler responseHandler = new TestSuccessResponseHandler();
         final Context context = new AttributesContext(new RootContext());
 
-        final FetchApiClientJwksFilter filter = new FetchApiClientJwksFilter(new ReturnsErrorsJwkStore());
+        final FetchApiClientJwksFilter filter = new FetchApiClientJwksFilter(new ReturnsExeptionsApiClientJwkSetService());
 
         final IllegalStateException exception = assertThrows(IllegalStateException.class,
                 () -> filter.filter(context, new Request(), responseHandler));
@@ -116,7 +97,7 @@ class FetchApiClientJwksFilterTest {
         final Context context = new AttributesContext(new RootContext());
         addApiClientToContext(context, new ApiClient());
 
-        final FetchApiClientJwksFilter filter = new FetchApiClientJwksFilter(new ReturnsErrorsJwkStore());
+        final FetchApiClientJwksFilter filter = new FetchApiClientJwksFilter(new ReturnsExeptionsApiClientJwkSetService());
 
         final IllegalStateException exception = assertThrows(IllegalStateException.class,
                 () -> filter.filter(context, new Request(), responseHandler));
@@ -124,126 +105,18 @@ class FetchApiClientJwksFilterTest {
     }
 
     @Test
-    void failsIfJwkSetServiceThrowsException() throws Exception {
-        final URL jwksUri = new URL("https://directory.com/jwks/12345");
-        final ApiClient apiClient = createApiClientWithJwksUri(jwksUri.toURI());
-        final TrustedDirectory trustedDirectory = new TrustedDirectoryOpenBankingTest();
+    void failsIfApiClientJwksSetServiceThrowsException() throws ExecutionException, InterruptedException, TimeoutException {
+        final FetchApiClientJwksFilter filter = new FetchApiClientJwksFilter(new ReturnsExeptionsApiClientJwkSetService());
 
-        final TestSuccessResponseHandler responseHandler = new TestSuccessResponseHandler();
         final Context context = new AttributesContext(new RootContext());
-        addApiClientToContext(context, apiClient);
-        addTrustedDirectoryToContext(context, trustedDirectory);
-
-        // Returns an Exception promise on every call
-        final JwkSetService jwkSetService = new ReturnsErrorsJwkStore();
-        final FetchApiClientJwksFilter filter = new FetchApiClientJwksFilter(jwkSetService);
-
-        final Promise<Response, NeverThrowsException> responsePromise = filter.filter(context, new Request(), responseHandler);
-        final Response response = responsePromise.get(1, TimeUnit.SECONDS);
-        assertEquals(Status.INTERNAL_SERVER_ERROR, response.getStatus());
-        assertFalse(responseHandler.hasBeenInteractedWith(), "ResponseHandler must not get invoked");
-    }
-
-    @Test
-    void failsIfJwksUriIsInvalid() throws Exception {
-        final ApiClient apiClient = createApiClientWithJwksUri(new URI("foo://bar"));
-        final TrustedDirectory trustedDirectory = new TrustedDirectoryOpenBankingTest();
-
-        final TestSuccessResponseHandler responseHandler = new TestSuccessResponseHandler();
-        final Context context = new AttributesContext(new RootContext());
-        addApiClientToContext(context, apiClient);
-        addTrustedDirectoryToContext(context, trustedDirectory);
-
-        final JwkSetService jwkSetService = new ReturnsErrorsJwkStore();
-        final FetchApiClientJwksFilter filter = new FetchApiClientJwksFilter(jwkSetService);
-
-        final Promise<Response, NeverThrowsException> responsePromise = filter.filter(context, new Request(), responseHandler);
-        final Response response = responsePromise.get(1, TimeUnit.SECONDS);
-        assertEquals(Status.INTERNAL_SERVER_ERROR, response.getStatus());
-        assertFalse(responseHandler.hasBeenInteractedWith(), "ResponseHandler must not get invoked");
-    }
-
-    @Test
-    void failsIfJwksUriIsNull() throws Exception {
-        final ApiClient apiClient = createApiClientWithJwksUri(null);
-        final TrustedDirectory trustedDirectory = new TrustedDirectoryOpenBankingTest();
-
-        final TestSuccessResponseHandler responseHandler = new TestSuccessResponseHandler();
-        final Context context = new AttributesContext(new RootContext());
-        addApiClientToContext(context, apiClient);
-        addTrustedDirectoryToContext(context, trustedDirectory);
-
-        final JwkSetService jwkSetService = new ReturnsErrorsJwkStore();
-        final FetchApiClientJwksFilter filter = new FetchApiClientJwksFilter(jwkSetService);
-
-        final Promise<Response, NeverThrowsException> responsePromise = filter.filter(context, new Request(), responseHandler);
-        final Response response = responsePromise.get(1, TimeUnit.SECONDS);
-        assertEquals(Status.INTERNAL_SERVER_ERROR, response.getStatus());
-        assertFalse(responseHandler.hasBeenInteractedWith(), "ResponseHandler must not get invoked");
-    }
-
-    @Test
-    void failsToGetJwksFromSoftwareStatementIfTrustedDirectorySoftwareStatementJwksClaimNameIsMissing() throws Exception {
-        final ReturnsErrorsJwkStore errorsJwkStore = new ReturnsErrorsJwkStore();
-        final FetchApiClientJwksFilter filter = new FetchApiClientJwksFilter(errorsJwkStore);
-        final JWKSet jwkSet = createJwkSet();
-        final URL secureApiGatewayJwksURI = new URL("https://blah.com");
-        final TrustedDirectory misconfiguredDirectory = new TrustedDirectorySecureApiGateway(secureApiGatewayJwksURI) {
-            @Override
-            public String getSoftwareStatementJwksClaimName() {
-                return null;
-            }
-        };
-        final ApiClient apiClient = createApiClientWithSoftwareStatementJwks(jwkSet,"jwks");
-        final Context context = new AttributesContext(new RootContext());
-        addApiClientToContext(context, apiClient);
-        addTrustedDirectoryToContext(context, misconfiguredDirectory);
-
+        addApiClientToContext(context, new ApiClient());
+        addTrustedDirectoryToContext(context, new TrustedDirectoryOpenBankingTest());
         final TestSuccessResponseHandler responseHandler = new TestSuccessResponseHandler();
         final Promise<Response, NeverThrowsException> responsePromise = filter.filter(context, new Request(), responseHandler);
-        final Response response = responsePromise.get(1, TimeUnit.SECONDS);
+
+        final Response response = responsePromise.get(1, TimeUnit.MILLISECONDS);
         assertEquals(Status.INTERNAL_SERVER_ERROR, response.getStatus());
-        assertFalse(responseHandler.hasBeenInteractedWith(), "ResponseHandler must not get invoked");
-    }
-
-    @Test
-    void failsToGetJwksFromSoftwareStatementIfClaimIsNull() throws Exception {
-        final ReturnsErrorsJwkStore errorsJwkStore = new ReturnsErrorsJwkStore();
-        final FetchApiClientJwksFilter filter = new FetchApiClientJwksFilter(errorsJwkStore);
-        final JWKSet jwkSet = createJwkSet();
-        final URL secureApiGatewayJwksURI = new URL("https://blah.com");
-        final TrustedDirectory misconfiguredDirectory = new TrustedDirectorySecureApiGateway(secureApiGatewayJwksURI);
-        final ApiClient apiClient = createApiClientWithSoftwareStatementJwks(jwkSet,null);
-        final Context context = new AttributesContext(new RootContext());
-        addApiClientToContext(context, apiClient);
-        addTrustedDirectoryToContext(context, misconfiguredDirectory);
-
-        final TestSuccessResponseHandler responseHandler = new TestSuccessResponseHandler();
-        final Promise<Response, NeverThrowsException> responsePromise = filter.filter(context, new Request(), responseHandler);
-        final Response response = responsePromise.get(1, TimeUnit.SECONDS);
-        assertEquals(Status.INTERNAL_SERVER_ERROR, response.getStatus());
-        assertFalse(responseHandler.hasBeenInteractedWith(), "ResponseHandler must not get invoked");
-    }
-
-    @Test
-    void failsToGetJwksFromSoftwareStatementIfClaimsIsInvalidJwksJson() throws Exception {
-        final ReturnsErrorsJwkStore errorsJwkStore = new ReturnsErrorsJwkStore();
-        final FetchApiClientJwksFilter filter = new FetchApiClientJwksFilter(errorsJwkStore);
-        final URL secureApiGatewayJwksURI = new URL("https://blah.com");
-        final TrustedDirectory misconfiguredDirectory = new TrustedDirectorySecureApiGateway(secureApiGatewayJwksURI);
-        final ApiClient apiClient = new ApiClient();
-        final JwtClaimsSet claimsSet = new JwtClaimsSet();
-        claimsSet.setClaim(misconfiguredDirectory.getSoftwareStatementJwksClaimName(), json(object(field("keys", "should be a list"))));
-        apiClient.setSoftwareStatementAssertion(new SignedJwt(new JwsHeader(), claimsSet, new byte[0], new byte[0]));
-        final Context context = new AttributesContext(new RootContext());
-        addApiClientToContext(context, apiClient);
-        addTrustedDirectoryToContext(context, misconfiguredDirectory);
-
-        final TestSuccessResponseHandler responseHandler = new TestSuccessResponseHandler();
-        final Promise<Response, NeverThrowsException> responsePromise = filter.filter(context, new Request(), responseHandler);
-        final Response response = responsePromise.get(1, TimeUnit.SECONDS);
-        assertEquals(Status.INTERNAL_SERVER_ERROR, response.getStatus());
-        assertFalse(responseHandler.hasBeenInteractedWith(), "ResponseHandler must not get invoked");
+        assertFalse(responseHandler.hasBeenInteractedWith()); // The filter must not have passed the request on to the next handler
     }
 
     @Nested
@@ -268,30 +141,40 @@ class FetchApiClientJwksFilterTest {
             final JsonValue config = json(object(field("jwkSetService", "JwkSetService")));
 
             final FetchApiClientJwksFilter filter = (FetchApiClientJwksFilter) new Heaplet().create(Name.of("test"), config, heap);
-            fetchJwkSetFromJwksUri(jwkSet, jwksUri, filter);
-            fetchJwkSetFromSoftwareStatement(filter);
+
+            final Context context = new AttributesContext(new RootContext());
+            addApiClientToContext(context, createApiClientWithJwksUri(jwksUri.toURI()));
+            addTrustedDirectoryToContext(context, new TrustedDirectoryOpenBankingTest());
+            final TestSuccessResponseHandler responseHandler = new TestSuccessResponseHandler();
+            final Promise<Response, NeverThrowsException> responsePromise = filter.filter(context, new Request(), responseHandler);
+            final Response response = responsePromise.get(1, TimeUnit.MILLISECONDS);
+            assertEquals(Status.OK, response.getStatus());
+            assertTrue(responseHandler.hasBeenInteractedWith());
+
+            // Verify JWKS is available in the context
+            assertEquals(jwkSet, FetchApiClientJwksFilter.getApiClientJwkSetFromContext(context));
         }
     }
 
-    private JWKSet createJwkSet() {
-        return new JWKSet(List.of(RestJwkSetServiceTest.createJWK(UUID.randomUUID().toString()),
-                                  RestJwkSetServiceTest.createJWK(UUID.randomUUID().toString())));
-    }
+    private static final class MockApiClientJwkSetService implements ApiClientJwkSetService {
 
-    private ApiClient createApiClientWithJwksUri(URI jwksUri) {
-        final ApiClient apiClient = new ApiClient();
-        apiClient.setJwksUri(jwksUri);
-        return apiClient;
-    }
+        private final JWKSet jwkSet;
 
-    private ApiClient createApiClientWithSoftwareStatementJwks(JWKSet jwkSet, String softwareStatementJwksClaimName) {
-        final ApiClient apiClient = new ApiClient();
-        final JwtClaimsSet claimsSet = new JwtClaimsSet();
-        if (softwareStatementJwksClaimName != null) {
-            claimsSet.setClaim(softwareStatementJwksClaimName, jwkSet.toJsonValue());
+        MockApiClientJwkSetService(JWKSet jwkSet) {
+            this.jwkSet = jwkSet;
         }
-        apiClient.setSoftwareStatementAssertion(new SignedJwt(new JwsHeader(), claimsSet, new byte[0], new byte[0]));
-        return apiClient;
+
+        @Override
+        public Promise<JWKSet, FailedToLoadJWKException> getJwkSet(ApiClient apiClient, TrustedDirectory trustedDirectory) {
+            return Promises.newResultPromise(jwkSet);
+        }
+    }
+
+    private static final class ReturnsExeptionsApiClientJwkSetService implements ApiClientJwkSetService {
+        @Override
+        public Promise<JWKSet, FailedToLoadJWKException> getJwkSet(ApiClient apiClient, TrustedDirectory trustedDirectory) {
+            return Promises.newExceptionPromise(new FailedToLoadJWKException("failed to load JWK"));
+        }
     }
 
     private void addApiClientToContext(Context context, ApiClient apiClient) {
@@ -303,20 +186,5 @@ class FetchApiClientJwksFilterTest {
                 .put(FetchTrustedDirectoryFilter.TRUSTED_DIRECTORY_ATTR_KEY, trustedDirectory);
     }
 
-    private void invokeFilterAndValidateSuccessResponse(JWKSet expectedJwkSet, ApiClient apiClient,
-            TrustedDirectory trustedDirectory, FetchApiClientJwksFilter filter) throws Exception {
-        final TestSuccessResponseHandler responseHandler = new TestSuccessResponseHandler();
-        final Context context = new AttributesContext(new RootContext());
-        addApiClientToContext(context, apiClient);
-        addTrustedDirectoryToContext(context, trustedDirectory);
 
-        assertNull(FetchApiClientJwksFilter.getApiClientJwkSetFromContext(context),
-                "there must be no apiClientJwkSet in the context before the filter is called");
-
-        final Promise<Response, NeverThrowsException> responsePromise = filter.filter(context, new Request(), responseHandler);
-        final Response response = responsePromise.get(1, TimeUnit.SECONDS);
-        assertEquals(Status.OK, response.getStatus());
-        assertTrue(responseHandler.hasBeenInteractedWith());
-        assertEquals(expectedJwkSet, FetchApiClientJwksFilter.getApiClientJwkSetFromContext(context));
-    }
 }
